requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1
docopt==0.6.2
idna==3.4
Pillow==9.4.0
pip==23.0.1
pipdeptree==2.6.0
pipreqs==0.4.11
requests==2.28.2
setuptools==65.6.3
sqlparse==0.4.3
tzdata==2021.1
urllib3==1.26.15
wheel==0.37.1
wincertstore==0.2
yarg==0.1.9
